fastDFS服务器 配置 上传图片 删除图片

+++ src/main/java/com/mx/minnong/controller/NewsController.java
@@ -8,14 +8,18 @@ import com.mx.minnong.pojo.vo.PageVo;
 import com.mx.minnong.service.NewsService;
 import com.mx.minnong.utils.ApiRootPath;
 import com.mx.minnong.utils.FastDFSClient;
+import com.mx.minnong.utils.FileTypeUtil;
 import com.mx.minnong.utils.JoeJSONResult;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.Date;
 import java.util.List;
 
@@ -32,22 +36,30 @@ public class NewsController {
     @Autowired
     private NewsService newsService;
 
+
+    @Autowired
+    private FastDFSClient fastDFSClient;
+
     /**
      * @param: [id]
      * 查询单条新闻
      */
-    @RequestMapping(value = "api/{id}",method = RequestMethod.GET)
-    public JoeJSONResult findOne(@PathVariable Integer id){
+    @RequestMapping(value = "api/{id}", method = RequestMethod.GET)
+    public JoeJSONResult findOne(@PathVariable Integer id) {
         News news = newsService.findOne(id);
-        return JoeJSONResult.ok(news);
+        if (news != null) {
+            return JoeJSONResult.ok(news);
+        } else {
+            return JoeJSONResult.errorMsg("没有这条新闻");
+        }
     }
 
     /**
      * @param: [pageNum pageSize]
      * 查询多条新闻  分页查询
      */
-    @RequestMapping(value = "api",method = RequestMethod.GET)
-    public JoeJSONResult findAll(PageVo pageVo){
+    @RequestMapping(value = "api", method = RequestMethod.GET)
+    public JoeJSONResult findAll(PageVo pageVo) {
         //TODO 新闻查询 生产环境redis缓存
         List<News> news = newsService.findAll(pageVo);
         PageInfo<News> pageInfo = new PageInfo<News>(news);
@@ -58,27 +70,51 @@ public class NewsController {
      * @param: [news]
      * 添加新闻需要权限 path > root
      */
-    @RequestMapping(value = ApiRootPath.ROOTPATH+"api",method = RequestMethod.POST)
-    public JoeJSONResult AddNews(News news){
-        //TODO 添加新闻  set 新闻发布人  上传图片
-        int result = newsService.saveNews(news);
+    @RequestMapping(value = ApiRootPath.ROOTPATH + "api", method = RequestMethod.POST)
+    public JoeJSONResult AddNews(News news, MultipartFile file) throws IOException {
+        //TODO set 新闻发布人
+        if (StringUtils.isBlank(file.getOriginalFilename())) {
+            return JoeJSONResult.errorMsg("新闻头图必须上传");
+        }
+        //判断是否图片格式 使用fastDFS服务器存储
+        if (FileTypeUtil.isImage(file.getOriginalFilename())) {
+            String path = fastDFSClient.uploadFile(file);
+            if (StringUtils.isNotBlank(path)) {
+                news.setNewsPicture(path);
+            }
+        }
+        newsService.saveNews(news);
         return JoeJSONResult.ok("添加成功");
     }
 
     /**
-     * @param: [id,news]
+     * @param: [id, news]
      * 修改新闻需要权限 path > root
      */
-    @RequestMapping(value = ApiRootPath.ROOTPATH+"api/{id}",method = RequestMethod.PUT)
-    public JoeJSONResult updateNews(@PathVariable Integer id,News news){
-        if (null != id && !"".equals(id)){
-            news.setNewsId(id);
-            System.out.println(new Date().toString());
-            news.setNewsUpdatetime(new Date());
-            int result = newsService.updateNews(news);
-            return JoeJSONResult.ok("修改成功");
-        }else {
-            log.error("【修改新闻】 id参数为空 news={}",news);
+    @RequestMapping(value = ApiRootPath.ROOTPATH + "api/{id}", method = RequestMethod.PUT)
+    public JoeJSONResult updateNews(@PathVariable Integer id, News news, MultipartFile file) throws IOException {
+        if (null != id && !"".equals(id)) {
+            News result = newsService.findOne(id);
+            //判断是否存在这条新闻
+            if (result != null) {
+                news.setNewsId(id);
+                //设置修改时间
+                news.setNewsUpdatetime(new Date());
+                //判断是否上传图片
+                if (StringUtils.isNotBlank(file.getOriginalFilename())) {
+                    //判断数据库是否已经存在图片
+                    if (StringUtils.isNotBlank(result.getNewsPicture())) {
+                        fastDFSClient.deleteFile(result.getNewsPicture());
+                    }
+                    news.setNewsPicture(fastDFSClient.uploadFile(file));
+                }
+                int i = newsService.updateNews(news);
+                return JoeJSONResult.ok("修改成功");
+            } else {
+                return JoeJSONResult.errorMsg("新闻不存在");
+            }
+        } else {
+            log.error("【修改新闻】 id参数为空 news={}", news);
             throw new NetException(NetEnum.ID_ISEMPTY);
         }
     }
@@ -87,10 +123,18 @@ public class NewsController {
      * @param: [id]
      * 删除新闻 如果新闻存在图片删除图片
      */
-    @RequestMapping(value = ApiRootPath.ROOTPATH+"api/{id}",method = RequestMethod.DELETE)
-    public JoeJSONResult deleteNews(@PathVariable Integer id){
-        //TODO 新闻图片删除
-        newsService.deleteNews(id);
+    @RequestMapping(value = ApiRootPath.ROOTPATH + "api/{id}", method = RequestMethod.DELETE)
+    public JoeJSONResult deleteNews(@PathVariable Integer id) {
+        News news = newsService.findOne(id);
+        if (news != null) {
+            //判断新闻是否存在图片  如果存在图片 则删除
+            if (StringUtils.isNotEmpty(news.getNewsPicture())) {
+                fastDFSClient.deleteFile(news.getNewsPicture());
+            }
+            newsService.deleteNews(id);
+        } else {
+            return JoeJSONResult.errorMsg("新闻不存在");
+        }
         return JoeJSONResult.ok("删除成功");
     }
 }
